refactor: Reduce queries for recording log in
Our original method for recording a log in used two queries because it
was easier to tell if we inserted a new record into the `users` table.
We actually want to look at if the user's name is empty to determine if
the user is "new". This allows us to use a single query.